Extract functions from cli
